requirements-dev.txt
# Development dependencies
flake8
pytest
torch
numpy
Pillow
transformers
torchvision
